requirements.txt
Pillow==8.0.0
pytest==6.2.2
numpy==1.21.0
